Raise errors from unexpected Metrc API responses (#1)
* raise full errors if there is a bad response from the API

when there is an unexpected result from the API, this gem should be able
to communicate that to the consumer. This is already done for a couple
of `get` requests, so why not handle it for any error type?

these errors are all grouped underneath Metrc::Errors::RequestError so
that they can be rescued from as a group.

error messages from a bad request are parsed for more detailed feedback.
Since the message response is not documented from a 500 I can't capture
that message with certainty.